+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Add GuardDuty service
 - Add AWS Macie
 - Adds EKS
+- Add AWS Pricing service
 
 ## [0.34.0] - 2018-09-05